+++ packages/cli/api-importers/v3-importer-commons/src/AbstractSpecConverter.ts
@@ -85,6 +85,7 @@ export abstract class AbstractSpecConverter<
                 hasStreamingEndpoints: false,
                 isAuthMandatory: true,
                 idempotencyKeyGeneration: undefined,
+                webhookSignatureVerification: undefined,
                 platformHeaders: {
                     language: "",
                     sdkName: "",